+++ Source/Scene/GlobeSurfaceTile.js
@@ -172,6 +172,45 @@ define([
         return undefined;
     };
 
+    GlobeSurfaceTile.prototype.pickTriangle = function(ray, mode, projection, cullBackFaces) {
+        var terrain = this.pickTerrain;
+        if (!defined(terrain)) {
+            return undefined;
+        }
+
+        var mesh = terrain.mesh;
+        if (!defined(mesh)) {
+            return undefined;
+        }
+
+        var vertices = mesh.vertices;
+        var indices = mesh.indices;
+        var encoding = mesh.encoding;
+
+        var length = indices.length;
+        for (var i = 0; i < length; i += 3) {
+            var i0 = indices[i];
+            var i1 = indices[i + 1];
+            var i2 = indices[i + 2];
+
+            var v0 = getPosition(encoding, mode, projection, vertices, i0, scratchV0);
+            var v1 = getPosition(encoding, mode, projection, vertices, i1, scratchV1);
+            var v2 = getPosition(encoding, mode, projection, vertices, i2, scratchV2);
+
+            var intersection = IntersectionTests.rayTriangle(ray, v0, v1, v2, cullBackFaces, scratchResult);
+            if (defined(intersection)) {
+                return {
+                    intersection : intersection,
+                    v0 : v0,
+                    v1 : v1,
+                    v2 : v2
+                };
+            }
+        }
+
+        return undefined;
+    };
+
     GlobeSurfaceTile.prototype.freeResources = function() {
         if (defined(this.waterMaskTexture)) {
             --this.waterMaskTexture.referenceCount;